Refactor assert200 and assert401 methods to utils helper
