+++ app/test/e2e/utils.js
@@ -240,7 +240,23 @@ const mockDeleteGraphAssocition = (id, application = 'rw', mockSuccess = true) =
         .reply(mockSuccess ? 200 : 404, { data: {} });
 };
 
+const assert401 = (response) => {
+    response.status.should.equal(401);
+    response.body.should.have.property('errors').and.be.an('array');
+    response.body.errors[0].should.have.property('detail').and.equal('Unauthorized');
+};
+
+const assert200 = (response, length = undefined) => {
+    response.status.should.equal(200);
+    response.body.should.have.property('data').and.be.an('array');
+    if (length) {
+        response.body.should.have.property('data').and.be.an('array').and.have.lengthOf(length);
+    }
+};
+
 module.exports = {
+    assert200,
+    assert401,
     createResource,
     createVocabulary,
     mockDataset,
