+++ app/test/e2e/vocabulary-find.spec.js
@@ -3,6 +3,8 @@ const chai = require('chai');
 const Resource = require('models/resource.model');
 const Vocabulary = require('models/vocabulary.model');
 
+const { assert200 } = require('./utils');
+
 const { getTestServer } = require('./test-server');
 
 chai.should();
@@ -25,27 +27,19 @@ describe('Vocabulary find test suite', () => {
     });
 
     it('Findind dataset vocabulary without auth returns 200 OK and a data array', async () => {
-        const response = await requester.get(`/api/v1/dataset/vocabulary/find`).send();
-        response.status.should.equal(200);
-        response.body.should.have.property('data').and.be.an('array');
+        assert200(await requester.get(`/api/v1/dataset/vocabulary/find`).send());
     });
 
     it('Findind widget vocabulary without auth returns 200 OK and a data array', async () => {
-        const response = await requester.get(`/api/v1/dataset/123/widget/vocabulary/find`).send();
-        response.status.should.equal(200);
-        response.body.should.have.property('data').and.be.an('array');
+        assert200(await requester.get(`/api/v1/dataset/123/widget/vocabulary/find`).send());
     });
 
     it('Findind layer vocabulary without auth returns 200 OK and a data array', async () => {
-        const response = await requester.get(`/api/v1/dataset/123/layer/vocabulary/find`).send();
-        response.status.should.equal(200);
-        response.body.should.have.property('data').and.be.an('array');
+        assert200(await requester.get(`/api/v1/dataset/123/layer/vocabulary/find`).send());
     });
 
     it('Findind all vocabulary without auth returns 200 OK and a data array', async () => {
-        const response = await requester.get(`/api/v1/vocabulary`).send();
-        response.status.should.equal(200);
-        response.body.should.have.property('data').and.be.an('array');
+        assert200(await requester.get(`/api/v1/vocabulary`).send());
     });
 
     afterEach(async () => {
